Add routes and methods for approving and disapproving host homes for hosts
The diff adds two new routes for approving and disapproving host homes for hosts. The corresponding methods have been added to the HostHomeController. The methods update the approval status of the host home and set flash messages for approval and disapproval. The view for the messages is returned.

Additionally, the diff updates the autoload\_classmap and autoload\_static files to include new classes and events related to chat functionality.