"Add job to send email to cohosts when a guest makes a booking request; update BookingsController to send booking request messages to cohosts"
This commit adds a new job, SendMailForChatToCohosts, which is used to send an email notification to cohosts when a guest makes a booking request. The BookingsController has been updated to use this job when sending messages to cohosts when a guest makes a booking request. Additionally, the code now checks if the host is a cohost and updates the host\_id accordingly. The commit also includes a small refactor to use the queue method instead of send for the NotificationMail in the handleBookingRequest method.